Use Longest Processing Time algorithm.
This is a simple and effective algorithm, which should out-perform what was being used previously. The former algorithm skipped over tests that it didn't think would fit, effectively saving some of the bigger tests for allocation at the end, and thereby not allowing the smaller tests to be used to "balance" each node. I suspect in a vast majority of cases, the longest processing time algorithm will result in faster overall execution time.